+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/Elements/CardSection/CardSectionElement.swift
@@ -26,7 +26,14 @@ final class CardSectionElement: ContainerElement {
     lazy var view: UIView = {
         #if !os(visionOS)
         if #available(iOS 13.0, macCatalyst 14, *), STPCardScanner.cardScanningAvailable {
-            return CardSectionWithScannerView(cardSectionView: cardSection.view, delegate: self, theme: theme, analyticsHelper: analyticsHelper, linkAppearance: linkAppearance)
+            return CardSectionWithScannerView(
+                cardSectionView: cardSection.view,
+                opensCardScannerAutomatically: opensCardScannerAutomatically,
+                delegate: self,
+                theme: theme,
+                analyticsHelper: analyticsHelper,
+                linkAppearance: linkAppearance
+            )
         } else {
             return cardSection.view
         }
@@ -37,6 +44,7 @@ final class CardSectionElement: ContainerElement {
     let cardSection: SectionElement
     let analyticsHelper: PaymentSheetAnalyticsHelper?
     let cardBrandFilter: CardBrandFilter
+    private let opensCardScannerAutomatically: Bool
 
     private let linkAppearance: LinkAppearance?
 
@@ -73,12 +81,14 @@ final class CardSectionElement: ContainerElement {
         theme: ElementsAppearance = .default,
         analyticsHelper: PaymentSheetAnalyticsHelper?,
         cardBrandFilter: CardBrandFilter = .default,
+        opensCardScannerAutomatically: Bool = false,
         linkAppearance: LinkAppearance? = nil
     ) {
         self.hostedSurface = hostedSurface
         self.theme = theme
         self.analyticsHelper = analyticsHelper
         self.cardBrandFilter = cardBrandFilter
+        self.opensCardScannerAutomatically = opensCardScannerAutomatically
         let nameElement = collectName
             ? PaymentMethodElementWrapper(
                 TextFieldElement.NameConfiguration(